refactor: re-strucured project folders.
- added a seperate folder for Layouts in 'src/components/shared' folder and placed all layouts in it.
- renamed 'Layout.jsx' to 'MainLayout.jsx'.
- fixed the import statements of related files.
