+++ src/components/shared/Layouts/AuthLayout/AuthLayout.css
@@ -1,10 +1,10 @@
 .mask-1,
 .mask-2 {
   mask:
-    url(../../assets/svg-shapes/rectangle.svg) no-repeat,
+    url(../../../../assets/svg-shapes/rectangle.svg) no-repeat,
     linear-gradient(#000 0 0);
   -webkit-mask:
-    url(../../assets/svg-shapes/rectangle.svg) no-repeat,
+    url(../../../../assets/svg-shapes/rectangle.svg) no-repeat,
     linear-gradient(#000 0 0);
   mask-size: 11rem;
   mask-composite: exclude;
