+++ src/components/shared/Layouts/MainLayout/MainLayout.jsx
@@ -1,17 +1,17 @@
 import { useState, useEffect } from 'react';
-import homePageBgImg from '../../../assets/images/backgrounds/home-page.jpg';
-import favoritesPageBgImg from '../../../assets/images/backgrounds/favorites-page.jpg';
-import playlistAndSubscriptionPageBgImg from '../../../assets/images/backgrounds/playlist and-subscription-page.jpg';
-import browsePageBgImg from '../../../assets/images/backgrounds/browse-page.jpg';
-import Header from '../Header/Header';
-import DesktopNavbar from '../DesktopNavbar/DesktopNavbar';
-import HamburgerMenu from '../HamburgerMenu/HamburgerMenu';
-import Player from '../Player/Player';
-import Footer from '../Footer/Footer';
-import Logo from '../../Logo/Logo';
+import homePageBgImg from '../../../../assets/images/backgrounds/home-page.jpg';
+import favoritesPageBgImg from '../../../../assets/images/backgrounds/favorites-page.jpg';
+import playlistAndSubscriptionPageBgImg from '../../../../assets/images/backgrounds/playlist and-subscription-page.jpg';
+import browsePageBgImg from '../../../../assets/images/backgrounds/browse-page.jpg';
+import Header from '../../Header/Header';
+import DesktopNavbar from '../../DesktopNavbar/DesktopNavbar';
+import HamburgerMenu from '../../HamburgerMenu/HamburgerMenu';
+import Player from '../../Player/Player';
+import Footer from '../../Footer/Footer';
+import Logo from '../../../Logo/Logo';
 import { useLocation, Outlet, Link } from 'react-router-dom';
 
-export default function Layout() {
+export default function MainLayout() {
   const [showDesktopLogoNavbar, setShowDesktopLogoNavbar] = useState(false);
   const currentPage = useLocation().pathname;
